Rename api.py to model.py. Fixes #91
