+++ sgkit/__init__.py
@@ -1,4 +1,6 @@
-from .api import (  # noqa: F401
+from .display import display_genotypes
+from .io.vcfzarr_reader import read_vcfzarr
+from .model import (  # noqa: F401
     DIM_ALLELE,
     DIM_PLOIDY,
     DIM_SAMPLE,
@@ -6,8 +8,6 @@ from .api import (  # noqa: F401
     create_genotype_call_dataset,
     create_genotype_dosage_dataset,
 )
-from .display import display_genotypes
-from .io.vcfzarr_reader import read_vcfzarr
 from .stats.aggregation import count_call_alleles, count_variant_alleles, variant_stats
 from .stats.association import gwas_linear_regression
 from .stats.hwe import hardy_weinberg_test
