+++ sgkit/model.py
